+++ Tests/BugsnagPerformanceTests/SpanOptionsTests.mm
@@ -48,6 +48,7 @@ using namespace bugsnag;
                                                                       startTime:SpanOptions().startTime 
                                                                      firstClass:BSGFirstClassNo
                                                             attributeCountLimit:128
+                                                      shouldInstrumentRendering:NO
                                                                    onSpanClosed:^(BugsnagPerformanceSpan * _Nonnull) {
     }];
     BugsnagPerformanceSpanOptions *objcOptions = [BugsnagPerformanceSpanOptions new];