+++ Tests/BugsnagPerformanceTests/SpanTests.mm
@@ -27,6 +27,7 @@ static BugsnagPerformanceSpan *spanWithStartTime(CFAbsoluteTime startTime, OnSpa
                                               startTime:startTime
                                              firstClass:BSGFirstClassUnset
                                     attributeCountLimit:128
+                              shouldInstrumentRendering:NO
                                            onSpanClosed:onEnded];
 }
 
@@ -276,6 +277,7 @@ static BugsnagPerformanceSpan *spanWithStartTime(CFAbsoluteTime startTime, OnSpa
                                                    startTime:0
                                                   firstClass:BSGFirstClassUnset
                                          attributeCountLimit:5
+                                   shouldInstrumentRendering:YES
                                                 onSpanClosed:^(BugsnagPerformanceSpan *) {}];
 
     // Note: "bugsnag.sampling.p" is automatically added.