+++ Tests/BugsnagPerformanceTests/WeakSpansListTests.mm
@@ -26,6 +26,7 @@ static BugsnagPerformanceSpan *createSpan() {
                                               startTime:SpanOptions().startTime 
                                              firstClass:BSGFirstClassNo
                                     attributeCountLimit:128
+                              shouldInstrumentRendering:NO
                                            onSpanClosed:^(BugsnagPerformanceSpan * _Nonnull) {
     }];
 }