+++ features/fixtures/ios/Scenarios/FrameMetricsSpanInstrumentRenderingOffScenario.swift
@@ -1,15 +1,15 @@
 //
-//  FrameMetricsSlowFramesScenario.swift
+//  FrameMetricsSpanInstrumentRenderingOffScenario.swift
 //  Fixture
 //
-//  Created by Robert B on 20/09/2024.
+//  Created by Robert B on 26/09/2024.
 //
 
 import Bugsnag
 import BugsnagPerformance
 
 @objcMembers
-class FrameMetricsSlowFramesScenario: Scenario {
+class FrameMetricsSpanInstrumentRenderingOffScenario: Scenario {
     
     override func configure() {
         super.configure()
@@ -17,7 +17,9 @@ class FrameMetricsSlowFramesScenario: Scenario {
     }
     
     override func run() {
-        let span = BugsnagPerformance.startSpan(name: "FrameMetricsSlowFramesScenario")
+        let options = BugsnagPerformanceSpanOptions()
+        options.setInstrumentRendering(false)
+        let span = BugsnagPerformance.startSpan(name: "FrameMetricsSpanInstrumentRenderingOffScenario", options: options)
         
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
             Thread.sleep(forTimeInterval: 0.3)