+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 
 ### Enhancements
 
+* Added rendering metrics to first class spans.
+  [319](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/319)
+
 * Added configurable limit to number of span attributes per span.
   [315](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/315)